% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colnames_type.R
\name{colnames_type}
\alias{colnames_type}
\title{colnames_type}
\usage{
colnames_type(df, type = "")
}
\arguments{
\item{df}{dataframe}

\item{type}{character, vector of types of columns to return}
}
\value{
character, vector column names corresponding to type
}
\description{
Given a dataframe and a character-vector of types, return the column names of that type
}
\details{
Following \code{dplyr::type_sum}, the most-common types are:

\describe{
  \item{lgl}{logical}
  \item{int}{integer}
  \item{dbl}{double (numeric)}
  \item{chr}{character}
  \item{fctr}{factor}
  \item{date}{date}
  \item{time}{time, such as \code{POSIXct}}
}
}
\examples{
colnames_type(xform_sample, c("int", "dbl"))
}

