% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/df_set_tz.R
\name{df_set_tz}
\alias{df_set_tz}
\title{df_set_tz}
\usage{
df_set_tz(df, tz = "UTC")
}
\arguments{
\item{df}{dataframe}

\item{tz}{timezone, an Olson timezone or "UTC" (default)}
}
\value{
\code{dplyr::tbl_df()}
}
\description{
Sets the timezone of all time-based columns in a dataframe
}
\examples{
df_set_tz(xform_sample, tz = "America/Chicago")
}

