% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_xform_sample.R
\docType{data}
\name{xform_sample}
\alias{xform_sample}
\title{dataset to demonstrate xform functions}
\format{A \code{"dplyr::tbl_df()"} with 10 rows and 5 variables:
\describe{
  \item{dtm_chicago}{POSIXct, timezone set to "America/Chicago"}
  \item{dbl_a}{double, sequence of doubles}
  \item{int_a}{integer, sequence of integers}
  \item{char_a}{character, sequence of character strings}
  \item{dtm_paris}{POSIXct, timezone set to "Europe/Paris"}
  \item{dbl_b}{double, sequence of doubles}
  \item{int_b}{integer, sequence of integers}
  \item{char_b}{character, sequence of character strings}
}}
\usage{
xform_sample
}
\description{
There is no particular significance to the data
}
\keyword{datasets}

